# [Core][Distributed] merge two broadcast_tensor_dict