… tests/worker/test_model_runner.py
… vllm/worker/neuron_model_runner.py
… vllm/worker/worker.py