We have two `broadcast_tensor_dict` call, one in worker, one in model_runner. We can merge them together in one `broadcast_tensor_dict` call.