@youkaichao 2024-06-09
@zhuohan123 PTAL, if you agree with this design, I can refactor the rest model runners as well.