@youkaichao 2024-06-10
the performance gain seems to be non-trivial:

with this PR:
Throughput: 31.08 requests/s, 15911.95 tokens/s
Throughput: 31.22 requests/s, 15986.09 tokens/s
Throughput: 31.18 requests/s, 15963.27 tokens/s

without this PR:
Throughput: 30.77 requests/s, 15752.00 tokens/s
Throughput: 30.58 requests/s, 15656.67 tokens/s
Throughput: 30.52 requests/s, 15626.47 tokens/s

script:

`python benchmarks/benchmark_throughput.py --output-len 256 --input 256 --model meta-llama/Llama-2-7b-hf -tp 8`

machine: 8*H100