@youkaichao 2024-06-10
> can we make the code more specific? Why do we design the function arguments with a general name like aux?

i feel this is quite general? the driver worker prepares input, separate it into objects to broadcast, and objects to keep for itself (i.e. the `aux`).