@zhuohan123 2024-06-10
But the method is called `execute_model`. I think my confusion is this: The function name is pretty specific (`execute_model`), but the arguments are very general: `broadcast_inputs` and `aux`.