@youkaichao 2024-06-10
How about this design:

```
seq_group_metadata_list  --prepare_inputs_to_broadcast--> tensor_dict_to_broadcast, auxiliary_data --convert_broadcast_inputs_to_model_input--> model_input --execute_model--> output
```

for non-driver workers:

```
broadcast_tensor_dict, None  --convert_broadcast_inputs_to_model_input--> model_input --execute_model--> output
```