@youkaichao 2024-06-10
> we can halve again by adding https://github.com/vllm-project/vllm/pull/4844 or equivalent

I'm planning to use share memory for transport, so that we don't need to broadcast metadata at all. Only tensors will be broadcasted.

> Spec decoding adds another per-step broadcast which I hope can be similarly coalesced with this one

Sure, we can do it in a followup PR.