@youkaichao 2024-06-26
close as https://github.com/vllm-project/vllm/pull/5408 is a superset of this pr.